+++ packages/snap/src/config.ts
@@ -21,19 +21,19 @@ export const config = {
   CIRCUIT_AUTH_FINAL_KEY_URL: './assets/circuits/auth/circuit_final.zkey',
 
   CIRCUIT_SIG_V2_ON_CHAIN_WASM_URL:
-    './assets/circuits/credentialAtomicQuerySigV2OnChain/circuit.wasm',
+    'https://storage.googleapis.com/rarimo-store/snap/QmS4vURQ1c8tgALSokdTYVqx5E9FmASbu964W3JevnM3B4',
   CIRCUIT_SIG_V2_ON_CHAIN_FINAL_KEY_URL:
-    './assets/circuits/credentialAtomicQuerySigV2OnChain/circuit_final.zkey',
+    'https://storage.googleapis.com/rarimo-store/snap/QmT45Y62hfZnADq6VvKGjNR8foNb2KjcyG4AStRRAN9iHm',
 
   CIRCUIT_SIG_V2_WASM_URL:
-    './assets/circuits/credentialAtomicQuerySigV2/circuit.wasm',
+    'https://storage.googleapis.com/rarimo-store/snap/QmYB5QLvH5ihiedxvzkG3XPQngjxcS8wc1xCAoKnGS5GfC',
   CIRCUIT_SIG_V2_FINAL_KEY_URL:
-    './assets/circuits/credentialAtomicQuerySigV2/circuit_final.zkey',
+    'https://storage.googleapis.com/rarimo-store/snap/QmeXxRXxYGCwa48ANikH5Knzi9cgkmhumPbMtjTKNYkThL',
 
   CIRCUIT_MTP_V2_WASM_URL:
-    './assets/circuits/credentialAtomicQueryMTPV2/circuit.wasm',
+    'https://storage.googleapis.com/rarimo-store/snap/QmRqGgnN6Qy4LuPxQKH2wrADNe4aJb8wYJhS1ky9zbLS8t',
   CIRCUIT_MTP_V2_FINAL_KEY_URL:
-    './assets/circuits/credentialAtomicQueryMTPV2/circuit_final.zkey',
+    'https://storage.googleapis.com/rarimo-store/snap/QmcLyDLPWJpyEWeR9KkWQuGHAqifnwpDAWBX1R6a7g6F6a',
 
   CIRCUIT_MTP_V2_ON_CHAIN_WASM_URL:
     './assets/circuits/credentialAtomicQueryMTPV2OnChain/circuit.wasm',